fix: rename folder hoc, hook
